server/requirements.txt
flask
janusgraphy
flask-restful
pyjwt